+++ app/src/main/java/com/bangkit/relaverse/ui/ViewModelFactory.kt
@@ -9,6 +9,7 @@ import com.bangkit.relaverse.ui.auth.AuthViewModel
 import com.bangkit.relaverse.ui.create_event.CreateEventViewModel
 import com.bangkit.relaverse.ui.main.DetailViewModel
 import com.bangkit.relaverse.ui.main.MainViewModel
+import com.bangkit.relaverse.ui.main.campaign.ListUserViewModel
 import com.bangkit.relaverse.ui.main.profile.ProfileViewModel
 
 class ViewModelFactory private constructor(
@@ -27,6 +28,8 @@ class ViewModelFactory private constructor(
             return CreateEventViewModel(repository) as T
         }else if (modelClass.isAssignableFrom(ProfileViewModel::class.java)) {
             return ProfileViewModel(repository) as T
+        }else if (modelClass.isAssignableFrom(ListUserViewModel::class.java)) {
+            return ListUserViewModel(repository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class: ${modelClass.name}")
     }
